+++ CHANGELOG.md
@@ -3,6 +3,16 @@
 -- TODO: Pending changes to documentation, under works.
 -- 
 
+# 0.9.85
+
+### Changes
+
+- Added new includeFile :file attribute 
+- Added new examples for file include
+- Run examples
+-- make sure you have installed http-server (npm install http-server -g) 
+-- run examples using npm (run examples and then visit http://127.0.0.1:8080/examples/ or http://127.0.0.1:8080/examples/inc-example.html
+
 # 0.9.83
 
 ### Changes
@@ -12,7 +22,7 @@
     - @+ for ^focus
     - ^- for ^blur
 - Added new examples
-- now one can install http-server (npm i http-server) and run the examples 
+- now one can install http-server (npm i http-server -g) and run the examples 
 
 
 # 0.9.82